apps/identity/pyproject.toml
[project]
name = "identity"
version = "0.1.0"
description = "身份认证微服务 - 公司管理与用户认证"
requires-python = ">=3.12"
dependencies = [
    "common",
    "fastapi>=0.115",
    "pydantic-settings>=2.1",
    "uvicorn>=0.32",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true

[tool.uv.sources]
common = { workspace = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
